+++ compiler/noirc_evaluator/src/ssa/parser/mod.rs
@@ -366,7 +366,7 @@ impl<'a> Parser<'a> {
 
         let value = self.parse_value_or_error()?;
         self.eat_or_error(Token::Keyword(Keyword::To))?;
-        let max_bit_size = self.eat_int_or_error()?.to_u128() as u32;
+        let max_bit_size = self.eat_int_or_error()?.to_u128() as u8;
         self.eat_or_error(Token::Keyword(Keyword::Bits))?;
         Ok(Some(ParsedInstruction::RangeCheck { value, max_bit_size }))
     }
@@ -494,12 +494,12 @@ impl<'a> Parser<'a> {
         if self.eat_keyword(Keyword::Truncate)? {
             let value = self.parse_value_or_error()?;
             self.eat_or_error(Token::Keyword(Keyword::To))?;
-            let bit_size = self.eat_int_or_error()?.to_u128() as u32;
+            let bit_size = self.eat_int_or_error()?.to_u128() as u8;
             self.eat_or_error(Token::Keyword(Keyword::Bits))?;
             self.eat_or_error(Token::Comma)?;
             self.eat_or_error(Token::Keyword(Keyword::MaxBitSize))?;
             self.eat_or_error(Token::Colon)?;
-            let max_bit_size = self.eat_int_or_error()?.to_u128() as u32;
+            let max_bit_size = self.eat_int_or_error()?.to_u128() as u8;
             return Ok(ParsedInstruction::Truncate { target, value, bit_size, max_bit_size });
         }
